+++ signup/index.html
@@ -61,7 +61,8 @@
       </div>
     </div>
 
-    <script type="module" src="../common.js"></script>
+    <script type="module" src="../js/auth.js"></script>
+    <script type="module" src="../js/uicontroller.js"></script>
     <script type="module" src="./app.js"></script>
   </body>
 </html>
